refactor evaluation panel to enable timestamp jumping from screen + clean imports
